move into directory and add selector tests
